requirements.txt
PyMuPDF
openpyxl
pandas
python-docx
openai
pdfplumber